+++ opengate/actors/miscactors.py
